Remove windows, osx impl for now
Cutting out the C++, this just reduces the surface area
for now, for that work. I primarily work on linux, so its
the one I care to carve time for. If/when time allows a C
version of Windows could be added back. Macos is so developer
hostile these days (Windows is too, but at least it has the
benefit of games needing to run there), I can't be bothered.